Move Parser into Io namespace and add tests
